% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{clean_asthma_control}
\alias{clean_asthma_control}
\title{Clean and derive asthma control measures from ISAR standardised asthma control dataset.}
\usage{
clean_asthma_control(df = asthma_control_labelled)
}
\arguments{
\item{df}{character. Name of labelled asthma control dataset.}
}
\description{
Recodes GINA asthma control responses to numeric values to calculate score and
uses to derive GINA asthma control classification (Well controlled vs
Partly controlled vs Uncontrolled). Derives categorical classifications for
asthma control based on available ACQ and ACT scores.
}
