% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{clean_biologics}
\alias{clean_biologics}
\title{Clean and derive measures from ISAR standardised biologics dataset.}
\usage{
clean_biologics(df = biologics_labelled)
}
\arguments{
\item{df}{character. Name of labelled biologics dataset.}
}
\description{
Creates additional short biologic name variable that excludes product names.
Creates biologic type variable that distinguishes between IL5 and IL5R based
on product. Renames variables to shorter prefix "bx_" in place of "bio_"
and expands start and end date variable names.
}
