% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{clean_biomarkers}
\alias{clean_biomarkers}
\title{Clean and derive biomarker measures from ISAR standardised biomarkers dataset.}
\usage{
clean_biomarkers(
  df = biomarkers_labelled,
  beos_range_low = 0,
  beos_range_high = 5000,
  feno_range_low = 0,
  feno_range_high = 300,
  ige_range_low = 0,
  ige_range_high = 100000,
  reduce_function = "max",
  drop_duplicate = TRUE,
  drop_rangefail = FALSE
)
}
\arguments{
\item{df}{character. Name of labelled biomarkers dataset.}

\item{beos_range_low}{numeric. Lowest plausible value for BEOS range check.}

\item{beos_range_high}{numeric. Highest plausible value for BEOS range check.}

\item{feno_range_low}{numeric. Lowest plausible value for FeNO range check.}

\item{feno_range_high}{numeric. Highest plausible value for FeNO range check.}

\item{ige_range_low}{numeric. Lowest plausible value for IgE range check.}

\item{ige_range_high}{numeric. Highest plausible value for Ige range check.}
}
\description{
Reshapes biomarkers dataset to long format with multiple rows per test.
Applies range checks (0 ≤ BEOS ≤ 5000; 0 ≤ FeNO ≤ 300; 0 ≤ IgE ≤ 100000)
to recorded biomarker measures. Flags cases Where multiple results are
recorded for a test on a single date, and allows for three approaches for
reducing these duplicates, either keeping minimum, mean, or maximum value.
Further cleaning steps to drop these duplicates and exclude measures failing
range checks can be specified.
}
