% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{clean_comorbidities}
\alias{clean_comorbidities}
\title{Clean and derive comorbidity measures from ISAR standardised comorbidities dataset.}
\usage{
clean_comorbidities(df = comorbidities_labelled)
}
\arguments{
\item{df}{character. Name of labelled comorbidity dataset.}
}
\description{
Clean and derive comorbidity measures from ISAR standardised comorbidities dataset.
}
