% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{clean_demographics}
\alias{clean_demographics}
\title{Clean and derive demographic measures from ISAR standardised demographics dataset.}
\usage{
clean_demographics(df = demographics_labelled)
}
\arguments{
\item{df}{character. Name of labelled demographics dataset.}
}
\description{
Derives asthma onset categories (<12, 12-40, >40 years), asthma duration (in years),
and asthma duration categories (<10, 10-<25, ≥25) from labelled demographic data.
}
