% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{clean_lifestyle}
\alias{clean_lifestyle}
\title{Clean and derive lifestyle measures from ISAR standardised lifestyle dataset.}
\usage{
clean_lifestyle(df = lifestyle_labelled)
}
\arguments{
\item{df}{character. Name of labelled demographics dataset.}
}
\description{
Applies range checks and plausibility checks height and weight measures and cleans
BMI measure accordingly. Derives categorical classifications for obese (BMI <30 vs. ≥30 kg/m2)
}
