% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{clean_spirometry}
\alias{clean_spirometry}
\title{Clean and derive spirometry measures from ISAR standardised spirometry dataset.}
\usage{
clean_spirometry(
  df = spirometry_labelled,
  fev1_range_low = 0.15,
  fev1_range_high = 8,
  fvc_range_low = 0.15,
  fvc_range_high = 8
)
}
\arguments{
\item{df}{character. Name of labelled spirometry dataset.}

\item{fev1_range_low}{numeric. Lowest plausible value for FEV1 range check.}

\item{fev1_range_high}{numeric. Highest plausible value for FEV1 range check.}

\item{fvc_range_low}{numeric. Lowest plausible value for FVC range check.}

\item{fvc_range_high}{numeric. Highest plausible value for FVC range check.}
}
\description{
Applies range checks and plausibility checks (FEV1 >= FVC) to pre- and
post-bronchodilator spirometry measures. Where post-bronchodilator measures
are available at every timepoint for a patient, they are prioritised over
pre-bronchodilator measures. Derives single composite FEV1, FVC, FEV1/FVC ratio,
and percent predicted FEV1 measures, along with categorical classifications
for FEV1 percent predicted (<80% vs ≥80%) and FEV1/FVC ratio (<0.70 vs ≥0.70).
}
