% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_datasets.R
\name{import_datasets}
\alias{import_datasets}
\title{Import ISAR standardised research datasets}
\usage{
import_datasets(filename, keep_raw = T)
}
\arguments{
\item{filename}{Filename of .csv as text}

\item{keep_raw}{Specify whether to keep "_raw" unformatted version of dataset object}
}
\value{
Assigns object in global environment using abbreviated dataset name
}
\description{
This function imports all listed datasets and creates a new "_labelled" version with all variables labelled, categorical variables formatted as factors with defined levels, and confirms and formats dates as defined in dictionary workbook
}
