% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_dataset.R
\name{output_dataset}
\alias{output_dataset}
\title{Save object as Rdata or Stata (version 14) and export dictionary in word or excel}
\usage{
output_dataset(
  df,
  export_dictionary = TRUE,
  excel_dictionary = FALSE,
  add_isar = TRUE,
  export_loc = "data_cleaning"
)
}
\arguments{
\item{df}{A data frame.}

\item{export_dictionary}{Logical. If `TRUE`, uses \link[labelled]{look_for}
to create a data dictionary object that can then be exported as Microsoft Excel
or Word (default) files.}

\item{excel_dictionary}{Logical. If `TRUE`, uses \link[writexl]{write_xlsx}
to export .xlsx file. If `FALSE` (default), uses \link[flextable]{save_as_docx}
to export .docx file.}

\item{add_isar}{Logical. If `TRUE`, adds "isar_" to front of file name.}

\item{export_loc}{Character vector. Specifies folder within project and
\link[here]{here} in which to save output files. Can also be full filepath.}
}
\description{
Processes specified dataframe to convert factors to labelled vectors using provided
helper function and \link[haven]{labelled}. Then saves dataframe as .dta
file using \link[haven]{write_dta}. The default is Stata version 14.
}
\examples{
output_to_stata(
   bl_biologic,
   export_dictionary = T,
   excel_dictionary = F,
   add_isar = F,
   export_loc = "data_cleaning"
)

}
