% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_empty_imputation_fields}
\alias{remove_empty_imputation_fields}
\title{Helper function to remove unused (empty) imputation fields from ISAR standardised datasets.}
\usage{
remove_empty_imputation_fields(df)
}
\description{
Helper function to remove unused (empty) imputation fields from ISAR standardised datasets.
}
