% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{clean_spirometry}
\alias{clean_spirometry}
\title{Clean ISAR standardised research datasets}
\usage{
clean_spirometry(
  df = spirometry_labelled,
  fev1_range_low = 0.15,
  fev1_range_high = 8,
  fvc_range_low = 0.15,
  fvc_range_high = 8
)
}
\arguments{
\item{df}{character. Name of labelled spirometry dataset.}

\item{fev1_range_low}{numeric. Lowest plausible value for FEV1 range check.}

\item{fev1_range_high}{numeric. Highest plausible value for FEV1 range check.}

\item{fvc_range_low}{numeric. Lowest plausible value for FVC range check.}

\item{fvc_range_high}{numeric. Highest plausible value for FVC range check.}
}
\description{
This function further processes standardised spirometry dataset to flag fev1 and fvc measures outside specified range, create single FEV1, FVC, FEV1/FVC, and percent predicted FEV1 prioritisint post-bronchodilator measures and excluding those failing range check and creates new categorical measures for FEV1, FEV1/FVC ratio.
}
